move topology header data fetching to react query
