+++ deepfence_frontend/apps/dashboard/src/queries/index.ts
@@ -7,6 +7,7 @@ import { postureQueries } from '@/queries/posture';
 import { registryQueries } from '@/queries/registry';
 import { searchQueries } from '@/queries/search';
 import { secretQueries } from '@/queries/secret';
+import { topologyQueries } from '@/queries/topology';
 import { vulnerabilityQueries } from '@/queries/vulnerability';
 
 export function invalidateQueries(queryKey: QueryKey) {
@@ -28,4 +29,5 @@ export const queries = mergeQueryKeys(
   malwareQueries,
   registryQueries,
   postureQueries,
+  topologyQueries,
 );
